Move locale macros to implementation
